+++ Syren.Server/Services/DistanceService.cs
@@ -7,7 +7,7 @@ namespace Syren.Server.Services;
 public class DistanceService : IDistanceService
 {
     private readonly Dictionary<uint, Speaker> _speakers = [];
-    private readonly Dictionary<uint, float> _speakerDistances = [];
+    private readonly Dictionary<uint, double> _speakerDistances = [];
     private uint _maxSpeakerId = 0;
 
     private readonly ILogger<DistanceService> _logger;
@@ -66,16 +66,16 @@ public class DistanceService : IDistanceService
         // Second speaker can be on any arbitrary point on the sphere
         // around the first at the distance between the two
         Speaker otherSpeaker = _speakers.First().Value;
-        float otherDistance = _speakerDistances[otherSpeaker.Id];
+        double otherDistance = _speakerDistances[otherSpeaker.Id];
 
-        return otherSpeaker.Position + new Vector3(otherDistance, 0.0f, 0.0f);
+        return otherSpeaker.Position + new Vector3((float)otherDistance, 0.0f, 0.0f);
     }
 
     private Vector3 GetAddedThirdSpeakerPosition()
     {
         // Third speaker can be on any arbitrary point on the circle
         // where the distances to both other speakers is correct
-        (Speaker Speaker, float Distance)[] speakers = _speakers
+        (Speaker Speaker, double Distance)[] speakers = _speakers
             .Take(2)
             .Select(keyValuePair =>
                 (keyValuePair.Value, _speakerDistances[keyValuePair.Key])
@@ -83,25 +83,25 @@ public class DistanceService : IDistanceService
 
         // Direction vector from the second speaker to the first
         Vector3 direction = speakers[1].Speaker.Position - speakers[0].Speaker.Position;
-        float twoSpeakerDistance = direction.Length();
+        double twoSpeakerDistance = direction.Length();
         direction = Vector3.Normalize(direction);
 
         // Points on the distance spheres farthest away from the other speaker's position
         Tuple<Vector3, Vector3> farPoints = new(
-            speakers[0].Speaker.Position + direction * speakers[0].Distance,
-            speakers[1].Speaker.Position - direction * speakers[1].Distance
+            speakers[0].Speaker.Position + direction * (float)speakers[0].Distance,
+            speakers[1].Speaker.Position - direction * (float)speakers[1].Distance
         );
         // Points on the distance spheres closest to the other speaker's position
         Tuple<Vector3, Vector3> nearPoints = new(
-            speakers[0].Speaker.Position - direction * speakers[0].Distance,
-            speakers[1].Speaker.Position + direction * speakers[1].Distance
+            speakers[0].Speaker.Position - direction * (float)speakers[0].Distance,
+            speakers[1].Speaker.Position + direction * (float)speakers[1].Distance
         );
 
         if (twoSpeakerDistance > speakers[0].Distance + speakers[1].Distance)
         {
             // Two speakers' distance spheres are disjoint
             // Place speaker on the connecting line at the right ratio
-            float distanceRatio = speakers[0].Distance / (speakers[0].Distance + speakers[1].Distance);
+            float distanceRatio = (float)(speakers[0].Distance / (speakers[0].Distance + speakers[1].Distance));
             return nearPoints.Item1 * distanceRatio + nearPoints.Item2 * (1.0f - distanceRatio);
         } else if (speakers[0].Distance > (speakers[0].Speaker.Position - farPoints.Item2).Length())
         {
@@ -117,7 +117,7 @@ public class DistanceService : IDistanceService
         {
             // Two distance spheres intersect
             // Place speaker at an intersection point
-            float speaker1Angle = MathF.Atan(speakers[1].Distance / speakers[0].Distance);
+            float speaker1Angle = MathF.Atan((float)(speakers[1].Distance / speakers[0].Distance));
             Vector3 orthogonal = Vector3.Normalize(new(
                 direction.Y + direction.Z,
                 direction.Z - direction.X,
@@ -188,19 +188,19 @@ public class DistanceService : IDistanceService
         return PoorMansGradientDescent(_speakers.Values.Center(), vector => MeanSquaredError(vector, spheres));
     }
 
-    private static float MeanSquaredError(Vector3 position, (Vector3 Center, float Radius)[] spheres)
+    private static double MeanSquaredError(Vector3 position, (Vector3 Center, double Radius)[] spheres)
     {
         return spheres.Select(sphere =>
         {
-            float distance = (sphere.Center - position).Length();
-            return (float)Math.Pow(sphere.Radius - distance, 2.0);
+            double distance = (sphere.Center - position).Length();
+            return Math.Pow(sphere.Radius - distance, 2.0);
         }).Sum() / spheres.Length;
     }
 
-    private Vector3 PoorMansGradientDescent(Vector3 initialValue, Func<Vector3, float> errorFunc)
+    private Vector3 PoorMansGradientDescent(Vector3 initialValue, Func<Vector3, double> errorFunc)
     {
         Vector3 result = initialValue;
-        float errorVal = errorFunc(result);
+        double errorVal = errorFunc(result);
 
         const double threshold = 0.3f;
         const uint maxIterations = 10;
@@ -211,9 +211,9 @@ public class DistanceService : IDistanceService
             if (errorVal <= threshold) return result;
 
             (result, errorVal) = _directions
-                .Select(direction => result + direction * errorVal)
+                .Select(direction => result + direction * (float)errorVal)
                 .Select(position => (position, errorFunc(position)))
-                .MinBy<(Vector3 Position, float ErrorVal), float>(tuple => tuple.ErrorVal);
+                .MinBy<(Vector3 Position, double ErrorVal), double>(tuple => tuple.ErrorVal);
         }
 
         _logger.LogWarning("User position {Position} has error value of {Error:0.00} > {Threshold:0.00}", result, errorVal, threshold);
